Force a sticky preference when resolving positions
Browsers like Safari will send ranges along with their beforeInput
events. These ranges have to be normalized in the same way as we
normalized the ranges we created ourselves during arrowleft and
arrowright, otherwise things look like they're out of sync.

When we decide to stick, we should also set a "resolve to the left" or
"resolve to the right" value. This value will allow us to resolve
positions at the end of nodes consistently, no matter how we get them.